+++ libs/base_core/src/code/util/_doc.rs
@@ -81,6 +81,25 @@ macro_rules! __doc {
             ),*
         )
     };
+    // WIP:
+    ( // Links to the module where this item was defined, in the devela crate.
+      // do NOT pass a leading slash.
+    location: $path:literal) => {
+        concat!("\n\n<sup title='Original location'>[`[", $path, "]`](",
+            $crate::doclink![custom devela $path @mod], ")</sup>\n\n")
+    };
+    ( // Links to the item in the module where it was defined, in the devela crate.
+    location_item: $path:literal) => {
+        concat!("\n\n<sup title='Original location'><em>[`", $path, "`](",
+            $crate::doclink![custom devela $path], ")</em></sup>\n\n")
+    };
+
+    ( // Links to the module where this item was defined, in the given crate.
+    location: $crate_id:ident, $path:literal) => {
+        concat!("\n\n<small title='Original location'>[`", $path, "`](",
+            $crate::doclink![custom $crate_id $path @mod], ")</small>\n\n")
+    };
+
     (
     // Shows the `Vendored` doc section and links to the info line.
     //