+++ libs/base_core/src/text/fmt/buf.rs
@@ -15,6 +15,8 @@ crate::_use! {compat::from_utf8}
 #[doc = crate::_TAG_FMT!()]
 /// Returns a formatted [`str`] slice backed by a buffer, non-allocating.
 ///
+#[doc = crate::_doc!(location: "text/fmt")]
+///
 /// Underneath it calls [`FmtWriter::format`] and [`format_args!`][crate::format_args].
 ///
 /// # Example