+++ libs/base_core/src/text/grapheme/nonul.rs
@@ -14,6 +14,8 @@ use crate::{IterChars, MismatchedCapacity, StringNonul, char7, char8, char16, do
 
 #[doc = crate::_TAG_TEXT!()]
 /// An <abbr title="Extended Grapheme Cluster">EGC</abbr> backed by a [`StringNonul`].
+///
+#[doc = crate::_doc!(location: "text")]
 #[must_use]
 #[repr(transparent)]
 #[derive(Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]