+++ libs/base_core/src/text/grapheme/u8.rs
@@ -13,6 +13,8 @@ use crate::{IterChars, MismatchedCapacity, StringU8, char7, char8, char16, unwra
 
 #[doc = crate::_TAG_TEXT!()]
 /// An <abbr title="Extended Grapheme Cluster">EGC</abbr> backed by a [`StringU8`].
+///
+#[doc = crate::_doc!(location: "text")]
 #[must_use]
 #[repr(transparent)]
 #[derive(Clone, PartialEq, Eq, Hash)]