+++ libs/base_core/src/text/str/str/mod.rs
@@ -26,6 +26,8 @@ use ::core::str::from_utf8_mut;
 #[doc = crate::_TAG_NAMESPACE!()]
 /// A string slice namespace.
 ///
+#[doc = crate::_doc!(location: "text/str")]
+///
 /// # Methods
 /// - [methods namespaced from `core::str`](#corestr-namespaced-methods)
 /// - [extra methods](#extra-methods)