+++ libs/base_core/src/text/str/u.rs
@@ -29,6 +29,8 @@ macro_rules! impl_str_u {
         #[doc = crate::_TAG_TEXT!()]
         #[doc = "A UTF-8–encoded string, backed by an array with [`" $t "::MAX`] bytes of capacity."]
         ///
+        #[doc = crate::_doc!(location: "text/str")]
+        ///
         #[doc = "Internally, the current length is stored as a [`" $t "`]."]
         ///
         /// ## Methods